Sync Direwolf V0.99: geometry optimization, and four convergence fixes
Adds geometry optimization and corrects four defects that were silently
limiting accuracy or wasting SCF iterations. Energies shift slightly for
every calculation - see the Bohr conversion note below.

New: geometry optimization
  runtype='opt' in &molecule, or an &opt block (every field optional),
  runs a quasi-Newton optimization with an RFO step and a trust radius.
  Redundant internal coordinates are the default, seeded by a diagonal
  model Hessian and back-transformed iteratively with dihedral
  wraparound handling; opt_coord='cart' selects Cartesians instead,
  seeded by a Lindh model Hessian. Octane B3LYP/6-31G takes 12 cycles in
  internals against 23 in Cartesians. The density is warm-started between
  steps, a trajectory is written to <base>.opt.xyz, and re-running the
  same input resumes from its last frame. The final geometry's result
  block has the same format as a single-point run's.

Fixed: COSX reused the first geometry's grid in multi-step runs
  The COSX build guard keyed only on cosx_ready and the basis dimension,
  neither of which changes when a driver calls the engine again for the
  same molecule at a new geometry, so every optimization step after the
  first used the starting geometry's grid, shell centres and
  overlap-fitting matrix. Single-point runs never saw this. A water
  RIJCOSX optimization went from failing at the 100-cycle cap to
  converging in 6.

Fixed: the DIIS Fock history was single precision
  The extrapolated Fock matrix inherits its history's precision directly,
  so real(4) storage put a ~3e-6 floor on the density change - above what
  scf_conv='fine' (1e-7) and 'tight' (1e-8) ask for, which meant those
  settings could never converge on their own criterion. Only the Fock
  history moves to real(8) (about 12 MB at 385 basis functions); the
  density and error-vector histories stay single precision, which
  reproduces the double-precision iteration count exactly. Ibuprofen
  r2SCAN-3c: 22 -> 18 iterations for a single point, and 13.7 -> 10.7
  SCF iterations per geometry step.

Fixed: the COSX incremental screening threshold was a constant
  An incremental exchange build screens a quantity whose scale is the
  density change itself, so a fixed threshold has a cliff: below some dP
  every shell pair falls under it, the exchange matrix stops being
  updated, and the accumulated error later discharges as a jump. The
  threshold now scales with dP, which removes the cliff instead of
  relocating it. Benzene D6h at scf_conv='tight': 37 -> 13 iterations.

Fixed: the Bohr conversion, in three ways
  The constant was the CODATA 1986 value, and the code disagreed with
  itself - the solvation cavity and the dispersion wrappers used a
  different one from the geometry. The literals in data/parameter.h also
  had no d0 suffix, making them single precision, PI included, which
  feeds every quadrature weight. And parameter.h is pulled in with
  INCLUDE, which the Makefile does not track, so editing it rebuilt
  nothing. All of it now matches pyscf's CODATA 2018 value: nuclear
  repulsion for a 42-atom molecule agrees with pyscf to 5e-9, where it
  was off by 8e-5 before - which had accounted for nearly all of a
  1e-4 Ha total-energy gap against both pyscf and ORCA.

Also: def2-SVP now covers Z=19-86 (it stopped at Ar, so no molecule
containing a transition metal could use it); the SCF no longer reports
convergence on an energy plateau while the density criterion is unmet;
and the README asks users to cite the methods Direwolf implements - COSX,
the Becke integration scheme, the -3c composites, D3/D4/gCP, VV10 and the
wB97M-V family, libxc and libcint.

Includes a workspace-size fix for the vendored multicharge: OpenBLAS's
sytrf query underestimates what the real call writes for some small
matrices, corrupting the heap past the work array.

+++ src/guess/mod_integrals_atomic_guess.f90
@@ -331,7 +331,7 @@ Feff = Hcore_l
 
 block
    real(8),allocatable :: X_l(:,:)
-   real(4),allocatable :: diis_Fa(:,:,:), diis_Fb(:,:,:)
+   real(8),allocatable :: diis_Fa(:,:,:), diis_Fb(:,:,:)
    real(4),allocatable :: diis_Pa(:,:,:), diis_Pb(:,:,:)
    real(4),allocatable :: diis_ea(:,:,:), diis_eb(:,:,:)
    integer :: DIIS_MAXl, diis_nl, multi_l
